Investigate for Unmanaged Profiling API. 02: Prepare the base process to replace a method body
Prepare the base process to replace a method body.
Add the base process to initialize the unmanaged profiling API.
Add the functions to operate SimpleHeap.
The 1st is Size method that returns the count of elements contained, the 2nd is operator[] method that accesses each element contained in itself.
Add the exception class NAnonymSystemException to handle system error.
Test ICorProfilerInfo2::SetEnterLeaveFunctionHooks2.
I will have to study in-line assembler...