Investigate for Unmanaged Profiling API. 02: Install the pseudo object
Install the pseudo object into the wrapper of the unmanaged meta data API and profiling API.
Because AppDomain** events were not notified when I performed this program actually. The cause of above was mscorlib.dll. I seem that this assembly is domain neutral.
Though I tried to make ProcessProfile that expresses the current process profiling info to manage the heap memories of AppDomain, Assembly and Module, the cyclic reference was occurred.
So I installed the pseudo object to manage them, for example, pseudo AppDomain to load some domain neutral assemblies such as mscorlib.dll.