Investigate for Unmanaged Profiling API. 03: Environment Preparation To Upgrade Visual Studio
Prepare my development environment to upgrade Visual Studio from 2008 to 2010.
In particular, corhlpr is reduced some classes from the version 2008, so some functions must be changed before the porting.

Add RapidVector that is similar to std::vector to manage small objects efficiently.

Upgrade all project files.
To be able to run all projects must spend a lot of time, so I modified only ProfilingApiSample03 to be runnable in the following environment:
- Windows 7 x64 (NOTE: The above projects were modified to be run as x86 application)
- Boost Ver.1.48.0
- Google Test Ver.1.6.0