Unmanaged Profiling API Sample 04: The Try Adding Result Of AppDomain Investigation Failed
I tried adding previous investigation result, but it failed :( Probably I seem that the cause is specialty of AppDomain which is belonged by mscorlib.
The AppDomain is neutral and has a special security permission. My CrossDomainDictionary can't ANY access from the AppDomain because CLR throws exceptions such as the following:
* System.IO.FileLoadException: "Loading this assembly would produce a different grant set from other instances. (Exception from HRESULT: 0x80131401)", if the first use is from default domain.
* System.ExecutionEngineException: -(No message), if the first use is from mscorlib's domain.

So, I change the mechanism to persist function pointers of managed methods after this commit.